fix for non numeric data types

+++ backtester/lookback_data.py
@@ -17,7 +17,7 @@ class LookbackDataEfficient:
         self.__endIndexList = 0
         self.__maxSize = self.computeMaxSize(lookbackSize, len(indexList))
         newEndIndexList = self.__endIndexList + self.__maxSize if (len(indexList) > self.__endIndexList + self.__maxSize) else len(self.__indexList)
-        self.__data = pd.DataFrame(columns=self.__columns, index=self.__indexList[self.__endIndexList:newEndIndexList], dtype=float)
+        self.__data = pd.DataFrame(columns=self.__columns, index=self.__indexList[self.__endIndexList:newEndIndexList])
         self.__endIndexList = newEndIndexList
 
     def computeMaxSize(self, lookbackSize, lenIndexList):
@@ -38,7 +38,7 @@ class LookbackDataEfficient:
             newEndIndexList = self.__endIndexList - self.__lookbackSize + self.__maxSize
             if newEndIndexList > len(self.__indexList):
                 newEndIndexList = len(self.__indexList)
-            newData = pd.DataFrame(columns=self.__columns, index=self.__indexList[self.__endIndexList - self.__lookbackSize: newEndIndexList], dtype=float)
+            newData = pd.DataFrame(columns=self.__columns, index=self.__indexList[self.__endIndexList - self.__lookbackSize: newEndIndexList])
             self.__endIndexList = newEndIndexList
             newData.iloc[0:self.__lookbackSize] = self.__data.iloc[self.__startLookbackData:self.__endLookbackData]
             self.__startLookbackData = 0
